Move DieselKeyStore up one level into diesel module
Moves `DieselKeyStore` into biome/key_management/store/diesel
and deletes biome/key_managment/store/diesel/postgres. This is to make
the file structure of key_managment consistent with the other store
modules. This change does not affect biome functionality.
